+++ packages/browser-wallet/src/background/identity-issuance.ts
@@ -77,7 +77,7 @@ export const identityIssuanceHandler: ExtensionMessageHandler = (msg) => {
             }
         }
         await openWindow();
-        bgMessageHandler.sendInternalMessage(InternalMessageType.EndIdentityIssuance, status);
+        bgMessageHandler.sendInternalMessage(InternalMessageType.EndIdentityIssuance, { ...response, status });
     };
 
     const { baseUrl, ...identityRequestInputs } = msg.payload;