+++ packages/browser-wallet/src/background/update.ts
@@ -1,4 +1,10 @@
-import { StorageAccessor, storedCredentials, storedIdentities } from '@shared/storage/access';
+import {
+    getGenesisHash,
+    StorageAccessor,
+    storedCredentials,
+    storedIdentities,
+    useIndexedStorage,
+} from '@shared/storage/access';
 import { Identity, WalletCredential } from '@shared/storage/types';
 
 const identityLock = 'Concordium_identity_lock';
@@ -27,7 +33,7 @@ async function editList<Type>(
     return navigator.locks.request(lock, async () => {
         const currentList = await storage.get();
         if (!currentList) {
-            return;
+            throw new Error('Attempt to edit non-existing list.');
         }
         const newList = [...currentList];
         for (const updated of edits) {
@@ -41,19 +47,19 @@ async function editList<Type>(
 }
 
 export async function addIdentity(identity: Identity | Identity[]): Promise<void> {
-    return addToList(identityLock, identity, storedIdentities);
+    return addToList(identityLock, identity, useIndexedStorage(storedIdentities, getGenesisHash));
 }
 
 export async function addCredential(cred: WalletCredential): Promise<void> {
-    return addToList(credentialLock, cred, storedCredentials);
+    return addToList(credentialLock, cred, useIndexedStorage(storedCredentials, getGenesisHash));
 }
 
 export function updateIdentities(updatedIdentities: Identity[]) {
     return editList(
         identityLock,
         updatedIdentities,
-        (identity) => (candidate) => identity.id === candidate.id,
-        storedIdentities
+        (identity) => (candidate) => identity.index === candidate.index,
+        useIndexedStorage(storedIdentities, getGenesisHash)
     );
 }
 
@@ -62,6 +68,6 @@ export function updateCredentials(updatedCredentials: WalletCredential[]) {
         credentialLock,
         updatedCredentials,
         (cred) => (candidate) => cred.credId === candidate.credId,
-        storedCredentials
+        useIndexedStorage(storedCredentials, getGenesisHash)
     );
 }