+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/i18n/da.ts
@@ -8,6 +8,8 @@ const t: typeof en = {
     blockHash: 'Blok hash ',
     events: 'Hændelser',
     rejectReason: 'Afvisningsårsag',
+    error: 'Fejl ved hentning af transaktionshistorik',
+    noTransactions: 'Denne konto har ingen transakationer endnu.',
 };
 
 export default t;