+++ packages/browser-wallet/src/popup/pages/NetworkSettings/i18n/da.ts
@@ -3,6 +3,12 @@ import type en from './en';
 const t: typeof en = {
     connect: 'Forbind',
     connected: 'Forbundet',
+    pendingIdentity: {
+        description:
+            'Du er ved at oprette en identitet, venligt færdiggør processen før du skifter dine networks indstillinger',
+        back: 'Gå tilbage',
+        abort: 'Afbryd Udstedelse',
+    },
 };
 
 export default t;