+++ packages/browser-wallet/src/popup/pages/NetworkSettings/i18n/en.ts
@@ -1,6 +1,12 @@
 const t = {
     connect: 'Connect',
     connected: 'Connected',
+    pendingIdentity: {
+        description:
+            'You are currently creating an identity, please finish the procedure before changing your network configuration',
+        back: 'Go back',
+        abort: 'Abort issuance',
+    },
 };
 
 export default t;