+++ packages/browser-wallet/src/popup/shell/i18n/locales/da.ts
@@ -7,13 +7,12 @@ import signMessage from '@popup/pages/SignMessage/i18n/da';
 import connectionRequest from '@popup/pages/ConnectionRequest/i18n/da';
 import settings from '@popup/pages/Settings/i18n/da';
 import networkSettings from '@popup/pages/NetworkSettings/i18n/da';
-import visualSettings from '@popup/pages/VisualSettings/i18n/da';
 import about from '@popup/pages/About/i18n/da';
 import addAccount from '@popup/pages/AddAccount/i18n/da';
 import identityIssuance from '@popup/pages/IdentityIssuance/i18n/da';
 import login from '@popup/pages/Login/i18n/da';
 import transactionLog from '@popup/pages/Account/TransactionLog/i18n/da';
-import changePasscode from '@popup/pages/ChangePasscode/i18n/en';
+import changePasscode from '@popup/pages/ChangePasscode/i18n/da';
 
 import type en from './en';
 
@@ -27,7 +26,6 @@ const t: typeof en = {
     connectionRequest,
     settings,
     networkSettings,
-    visualSettings,
     addAccount,
     identityIssuance,
     about,