+++ packages/browser-wallet/src/popup/store/identity.ts
@@ -3,25 +3,25 @@ import { atom } from 'jotai';
 import { selectAtom } from 'jotai/utils';
 import { atomWithChromeStorage } from './utils';
 
-export const identitiesAtom = atomWithChromeStorage<Identity[]>(ChromeStorageKey.Identities, []);
+export const identitiesAtom = atomWithChromeStorage<Identity[]>(ChromeStorageKey.Identities, [], false, true);
 export const pendingIdentityAtom = atomWithChromeStorage<Omit<PendingIdentity, 'location'> | undefined>(
     ChromeStorageKey.PendingIdentity,
     undefined
 );
-export const selectedIdentityIdAtom = atomWithChromeStorage<number>(ChromeStorageKey.SelectedIdentity, 0);
+export const selectedIdentityIndexAtom = atomWithChromeStorage<number>(ChromeStorageKey.SelectedIdentity, 0);
 export const selectedIdentityAtom = atom<Identity | undefined, Identity | undefined>(
     (get) => {
         const identities = get(identitiesAtom);
-        const id = get(selectedIdentityIdAtom);
-        return identities.find((a) => a.id === id);
+        const selectedIndex = get(selectedIdentityIndexAtom);
+        return identities.find((a) => a.index === selectedIndex);
     },
     (get, set, selectedIdentity) => {
         // Also update the identities atom.
         if (selectedIdentity) {
             const identities = get(identitiesAtom);
-            const id = get(selectedIdentityIdAtom);
-            const index = identities.findIndex((i) => i.id === id);
-            if (selectedIdentity.id !== id) {
+            const selectedIndex = get(selectedIdentityIndexAtom);
+            const index = identities.findIndex((i) => i.index === selectedIndex);
+            if (selectedIdentity.index !== selectedIndex) {
                 throw new Error('Updating selected identity with new id');
             }
             const newIdentities = [...identities];
@@ -36,7 +36,7 @@ export const identityProvidersAtom = atomWithChromeStorage<IdentityProvider[]>(C
 export const identityNamesAtom = selectAtom(identitiesAtom, (identities) => {
     const map = {} as Record<number, string>;
     identities.forEach((identity) => {
-        map[identity.id] = identity.name;
+        map[identity.index] = identity.name;
     });
     return map;
 });